requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
pyyaml>=6.0
pydantic>=2.0
pydantic-settings>=2.0
jinja2>=3.1
python-multipart>=0.0.9
aiosqlite>=0.20.0
